Изменения после комментариев по ревью №2

+++ src/test/java/tests/TestCourierLogin.java
@@ -1,3 +1,9 @@
+package tests;
+
+import apiHelper.CourierCreateApi;
+import apiHelper.CourierLoginApi;
+import apiHelper.Courier;
+import apiHelper.CourierApi;
 import com.google.gson.Gson;
 import io.qameta.allure.Description;
 import io.qameta.allure.Step;
@@ -12,20 +18,24 @@ public class TestCourierLogin {
 
     private Courier courier;
     private CourierApi CourierApi;
+    private boolean isCourierCreated;
     private final Gson gson = new Gson();
 
     @Before
     public void setUp() {
         // Создаем курьера перед каждым тестом
         CourierApi = new CourierApi();
-        courier = CourierApi.createCourier();
-        CourierApi.createCourierRequest(courier); // Создаем курьера через API
+        courier = CourierCreateApi.createCourier();
+        //CourierApi.createCourierRequest(courier);
+        isCourierCreated = false;
     }
 
     @After
     public void tearDown() {
         // Удаляем курьера после теста
-        CourierApi.deleteCourier(courier.getLogin(), courier.getPassword());
+        if (isCourierCreated) {
+            CourierApi.deleteCourier(courier.getLogin(), courier.getPassword());
+        }
     }
 
     @Test
@@ -33,12 +43,14 @@ public class TestCourierLogin {
     @Step("Возможность логина")
     public void courierCanLogin() {
         // Создаем объект для логина
-        CourierApi.LoginRequest loginRequest = new CourierApi.LoginRequest(courier.getLogin(), courier.getPassword());
-        Response response = CourierApi.doPostRequest(CourierApi.apiPathLogin, loginRequest);
+        CourierApi.createCourierRequest(courier); //создаем и логинимся курьером только в этом тесте
+        CourierLoginApi.LoginRequest loginRequest = new CourierLoginApi.LoginRequest(courier.getLogin(), courier.getPassword());
+        Response response = CourierApi.doPostRequest(CourierApi.API_LOGIN_PATH, loginRequest);
 
         response.then()
                 .statusCode(200) // Успешный запрос
                 .body("id", notNullValue()); // Проверка, что возвращается id
+        isCourierCreated = true;
     }
     @Test
     @Description("Проверка не валидных данных")
@@ -47,9 +59,9 @@ public class TestCourierLogin {
         String login = "invalidLogin";
         String password = "invalidPassword";
 
-        CourierApi.LoginRequest invalidLoginRequest = new CourierApi.LoginRequest(login, password);
+        CourierLoginApi.LoginRequest invalidLoginRequest = new CourierLoginApi.LoginRequest(login, password);
 
-        Response response = CourierApi.doPostRequest(CourierApi.apiPathLogin, invalidLoginRequest);
+        Response response = CourierApi.doPostRequest(CourierApi.API_LOGIN_PATH, invalidLoginRequest);
 
         response.then()
                 .statusCode(404) // Код ошибки для неверных учетных данных
@@ -59,9 +71,9 @@ public class TestCourierLogin {
     @Description("Проверка пустых полей")
     @Step("Логин с пустым логином")
     public void loginWithoutRequiredFields() {
-        CourierApi.LoginRequest loginRequest = new CourierApi.LoginRequest("", "somePassword");
+        CourierLoginApi.LoginRequest loginRequest = new CourierLoginApi.LoginRequest("", "somePassword");
 
-        Response response = CourierApi.doPostRequest(CourierApi.apiPathLogin, loginRequest);
+        Response response = CourierApi.doPostRequest(CourierApi.API_LOGIN_PATH, loginRequest);
 
         response.then()
                 .statusCode(400) // Код ошибки для отсутствующих обязательных полей
